Entity placement, destruction, and polish

+++ TunnelRats.gmx/scripts/scr_EntityBreak.gml
@@ -1,38 +1,24 @@
-/// scr_EntityBreak(entityInstance, chunk,x, y, z);
+/// scr_EntityBreak(entityInstance, chunk,x, y, z, doDrops);
 // test delete block!
 var breaking = argument0.EntityData;
 var chunk = argument1;
 var xx = argument2;
 var yy = argument3;
 var zz = argument4;
+var doDrops = argument5;
 
 // check if not air than do look checks
 if(breaking != -1)
 {
-    // create a queue of entities to spawn!
-    var spawnQueue = ds_queue_create();
-    
-    var blockDefinition = scr_EntityGetLibraryDef( breaking[? "DataName"] + ":" + breaking[? "Name"]);
-    
-    // standard breaking
-    var dropTable = ds_map_find_value(blockDefinition, "DropsOnBreak");
-    if(dropTable != -1)
+    if(doDrops)
     {
-        for (var i=0; i<ds_list_size(dropTable); i+=1)
-        {
-            var dropEntry = dropTable[| i];
-            
-            if(random(1) < dropEntry[| 1])
-            {
-                ds_queue_enqueue(spawnQueue,dropEntry[| 0]);
-            }
-        }
-    }
-    
-    // first time breaking has some unique drops!
-    if(breaking[? "FirstSpawn"])
-    {
-        var dropTable = ds_map_find_value(blockDefinition, "ExtraDropOnFirstBreak");
+        // create a queue of entities to spawn!
+        var spawnQueue = ds_queue_create();
+        
+        var blockDefinition = scr_EntityGetLibraryDef( breaking[? "DataName"] + ":" + breaking[? "Name"]);
+        
+        // standard breaking
+        var dropTable = ds_map_find_value(blockDefinition, "DropsOnBreak");
         if(dropTable != -1)
         {
             for (var i=0; i<ds_list_size(dropTable); i+=1)
@@ -45,26 +31,44 @@ if(breaking != -1)
                 }
             }
         }
-    }
-
-
-    while (ds_queue_size(spawnQueue) > 0)
-    {
-        // extract data
-        var newEntName = ds_queue_dequeue(spawnQueue);
-        var entityDat = scr_EntityInitData(newEntName,xx,yy,zz);
-        entityDat[? "InternalCount"] = 1; // number in each stack!
         
-        // spawn breaking entities
-        scr_EntityRealizeInstance( entityDat,chunk,-1, true);
-    }
+        // first time breaking has some unique drops!
+        if(breaking[? "FirstSpawn"])
+        {
+            var dropTable = ds_map_find_value(blockDefinition, "ExtraDropOnFirstBreak");
+            if(dropTable != -1)
+            {
+                for (var i=0; i<ds_list_size(dropTable); i+=1)
+                {
+                    var dropEntry = dropTable[| i];
+                    
+                    if(random(1) < dropEntry[| 1])
+                    {
+                        ds_queue_enqueue(spawnQueue,dropEntry[| 0]);
+                    }
+                }
+            }
+        }
     
-    ds_queue_destroy(spawnQueue);
+    
+        while (ds_queue_size(spawnQueue) > 0)
+        {
+            // extract data
+            var newEntName = ds_queue_dequeue(spawnQueue);
+            var entityDat = scr_EntityInitData(newEntName,xx,yy,zz);
+            entityDat[? "InternalCount"] = 1; // number in each stack!
+            
+            // spawn breaking entities
+            scr_EntityRealizeInstance( entityDat, chunk, noone, true);
+        }
+        
+        ds_queue_destroy(spawnQueue);
+    }
 }
 
 // spawn post destruction entity!
 var entityDat = scr_EntityInitData(breaking[? "BecomesOnBreak"], xx, yy, zz);
-var newEntityInstance = scr_EntityRealizeInstance( entityDat,chunk,-1, false);
+var newEntityInstance = scr_EntityRealizeInstance( entityDat, chunk, noone, false);
 
 // remove the original entity!
 with argument0 instance_destroy();
